# libs/api-tosspaper/openapi.yaml
openapi: 3.1.0
servers:
  - url: http://localhost:8080
    description: Development server
  - url: http://localhost:8080
    description: Local development server
info:
  title: Purchase Order System API
  version: v1
  description: REST API for creating resources in the Purchase Order System.
  contact:
    name: Acme Corp
    url: https://localhost:8080
components:
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          example: "Acme Analytics Platform"
        message:
          type: string
          example: "An error occurred while processing your request"

    # Sender Approval Request
...
    IntegrationSettings:
      type: object
      properties:
        currency:
          type: string
          description: "Company currency (ISO 4217)"
          example: "CAD"
        autoApprovalEnabled:
          type: boolean
          description: "Whether auto-approval is enabled"
          example: true
        autoApprovalThreshold:
          type: number
          format: decimal
          description: "Documents below this amount are auto-approved"
          example: 5000.00
    IntegrationSettingsUpdate:
      type: object
      properties:
        currency:
          type: string
          description: "Company currency (ISO 4217)"
          example: "CAD"
        autoApprovalEnabled:
          type: boolean
          description: "Whether auto-approval is enabled"
        autoApprovalThreshold:
          type: number
          format: decimal
          description: "Documents below this amount are auto-approved"
    IntegrationConnection:
      type: object
      properties:
        id:
          type: string
          description: "Connection ID"
        provider:
          type: string
          enum: [QUICKBOOKS, XERO, SAGE]
          description: "Integration provider"
        status:
          type: string
          enum: [disabled, enabled, expired, revoked]
          description: "Connection status. disabled = sync off, enabled = sync on, expired = token expired, revoked = user disconnected"
        externalCompanyName:
          type: string
          description: "Company name in external system"
        realmId:
          type: string
          description: "External realm/company ID"
        lastSyncAt:
          type: string
          format: date-time
          description: "Last sync timestamp"
        connectedAt:
          type: string
          format: date-time
          description: "When the connection was established"
        preferences:
          $ref: '#/components/schemas/IntegrationPreferences'
          nullable: true
          description: "Integration preferences (default currency, multicurrency status)"
    IntegrationPreferences:
      type: object
      properties:
        defaultCurrency:
          type: string
          nullable: true
          description: "ISO 4217 currency code for the default currency from provider (e.g., USD, CAD)"
          example: "USD"
        multicurrencyEnabled:
          type: boolean
          nullable: true
          description: "Whether multicurrency is enabled in the external provider"
          example: true
    IntegrationAuthUrl:
      type: object
      properties:
        authUrl:
          type: string
          format: uri
          description: "OAuth authorization URL"
        state:
          type: string
          description: "OAuth state for CSRF protection"
    IntegrationProvider:
      type: object
      properties:
        id:
          type: string
          description: "Provider ID (e.g., quickbooks, xero, sage)"
          example: "quickbooks"
        displayName:
          type: string
          description: "Display name for the provider"
          example: "QuickBooks Online"
        category:
          type: string
          description: "Category of integration (e.g., accounting)"
          example: "accounting"
  responses:
    BadRequest:
      description: Bad Request - The request could not be understood by the server due to malformed syntax.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized - Authentication credentials are required and have failed or have not yet been provided.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden - The server understood the request, but is refusing to fulfill it.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Not Found - The requested resource could not be found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
security:
  - jwtAuth: []
paths:
...
  /v1/companies/{companyId}/integration-settings:
    get:
      operationId: getIntegrationSettings
      tags:
        - "Integrations"
      summary: "Get company integration settings"
      description: "Retrieves integration settings for a company including auto-approval configuration."
      parameters:
        - name: X-Context-Id
          in: header
          required: true
          description: Company ID
          schema:
            type: string
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: "Company ID"
      responses:
        '200':
          description: "Integration settings"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntegrationSettings'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      operationId: updateIntegrationSettings
      tags:
        - "Integrations"
      summary: "Update company integration settings"
      description: "Updates integration settings for a company."
      parameters:
        - name: X-Context-Id
          in: header
          required: true
          description: Company ID
          schema:
            type: string
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: "Company ID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IntegrationSettingsUpdate'
      responses:
        '200':
          description: "Updated integration settings"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntegrationSettings'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # Integration Connections
  /v1/integrations/connections:
    get:
      operationId: listIntegrationConnections
      tags:
        - "Integrations"
      summary: "List integration connections"
      description: "Lists all integration connections for a company."
      parameters:
        - name: X-Context-Id
          in: header
          required: true
          description: Company ID
          schema:
            type: string
      responses:
        '200':
          description: "List of connections"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/IntegrationConnection'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /v1/integrations/connections/{connectionId}:
    put:
      operationId: updateIntegrationConnection
      tags:
        - "Integrations"
      summary: "Update integration connection"
      description: "Update connection status (enable/disable sync). Only 'enabled' and 'disabled' status values are allowed via API."
      parameters:
        - name: X-Context-Id
          in: header
          required: true
          description: Company ID
          schema:
            type: string
        - name: connectionId
          in: path
          required: true
          schema:
            type: string
          description: "Connection ID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  type: string
                  enum: [enabled, disabled]
                  description: "Connection status. 'enabled' = sync on, 'disabled' = sync off"
      responses:
        '200':
          description: "Connection updated"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntegrationConnection'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: disconnectIntegration
      tags:
        - "Integrations"
      summary: "Disconnect an integration"
      description: "Disconnects and removes an integration connection."
      parameters:
        - name: X-Context-Id
          in: header
          required: true
          description: Company ID
          schema:
            type: string
        - name: connectionId
          in: path
          required: true
          schema:
            type: string
          description: "Connection ID"
      responses:
        '204':
          description: "Integration disconnected"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # Integration Providers
  /v1/integrations/providers:
    get:
      operationId: getIntegrationProviders
      tags:
        - "Integrations"
      summary: "Get available integration providers"
      description: "Returns list of available integration providers (QuickBooks, Xero, Sage, etc.)."
      responses:
        '200':
          description: "List of available providers"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/IntegrationProvider'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # OAuth Flow
  /v1/integrations/{providerId}/auth-urls:
    post:
      operationId: createAuthUrl
      tags:
        - "Integrations"
      summary: "Create OAuth authorization URL"
      description: "Creates a new OAuth authorization request and returns the authorization URL. Each call generates a new state token."
      parameters:
        - name: X-Context-Id
          in: header
          required: true
          description: Company ID
          schema:
            type: string
        - name: providerId
          in: path
          required: true
          schema:
            type: string
          description: "Provider ID (e.g., quickbooks, xero, sage)"
      responses:
        '200':
          description: "OAuth authorization URL"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntegrationAuthUrl'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # Items
...